Re-create resources that have been deleted by hand
Whenever the API returns a 404 during a read it means that the resource
was deleted by hand through the UI. The expected behaviour after this
happens is that the terraform provider should try to re-created the
resource that was deleted.
To implement this and touch the least amount of the code we added a
wrapper that receives a ReadFunc and handles the API error that we care
about.

Closes #101

Signed-off-by: Matias Pan <matias.pan26@gmail.com>