Update integration name restrictions
The length limit is not explicitly stated on Opsgenie's documentation,
but when you try creating a resource with more than 250 characters this
is what the API returns:

    Error: Error occurred with Status code: 422, Message: [Integration name] length exceeds the maximum value [250], Took: 0.021000, RequestId: xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx

The format doesn't seem restricted at all, and you can even add emojis
to the integration name. Since the custom validation function would end
up being just a check for the name length, it was replaced with a call
to the validation helpers in each integration.