fix(resource_opsgenie_team): Sort team members
This fixes a bug where team members were returned from the provider in
an order that is inconsistent with that which is stored in state. This
resulted in perpetual drift, where team members were always shuffled
around on every plan.

This was fixed by using the same sorting algorithm each time these team
members are read.